pyproject.toml
[project]
name = "mechanistic-interpretability-case-study"
version = "0.2.0"
description = "Decomposing Contextual Embeddings: Mechanistic Interpretability via Sparse Autoencoders"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "clearml>=2.1.5",
    "datasets>=4.8.4",
    "einops>=0.8.2",
    "python-dotenv>=1.2.2",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
    "transformer-lens>=2.18.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.10",
]
